Version 21 09 2026
Fin déclaration Constraints Business données lambdas.

+++ EconomicTrendsPolLESSIRE.Application/Interfaces/IMarketTradeService.cs
@@ -6,12 +6,12 @@ namespace EconomicTrendsPolLESSIRE.Application.Interfaces
 {
     public interface IMarketTradeService
     {
-        Task<MarketTradeDTO?> GetByIdAsync(int id);
+        Task<MarketTradeDTO?> GetByIdAsync(long id);
         Task<MarketTradeDTO?> SaveAsync(MarketTradeDTO dto);
         Task<MarketTrade?> SaveMarketTradeAsync(MarketTrade markettrd, CancellationToken ct = default);
         Task<IEnumerable<MarketTrade>> GetAllAsync(int limit = 500, CancellationToken ct = default);
-        Task<MarketTrade?> GetMarketTradeByIdAsync(int id, CancellationToken ct = default);
-        Task<bool> DeleteMarketTradeAsync(int id, CancellationToken ct = default);
+        Task<MarketTrade?> GetMarketTradeByIdAsync(long id, CancellationToken ct = default);
+        Task<bool> DeleteMarketTradeAsync(long id, CancellationToken ct = default);
     }
 }
